requirements.txt
# IMPORTANT: This project uses different requirements files:
# - requirements_docker.txt: For Docker training environment  
# - requirements_data.txt: For local data download environment
# Please use the appropriate file based on your needs

# Core ML libraries (PyTorch is installed separately in Dockerfile)
# torch==2.1.0
# torchvision==0.16.0
# torchaudio==2.1.0

# Diffusion models
diffusers==0.24.0
transformers==4.36.0
accelerate==0.25.0
safetensors==0.4.1

# LoRA and parameter-efficient fine-tuning
peft==0.7.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
audioread==3.0.1

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Data handling
numpy==1.24.4
pandas==2.1.4
scipy==1.11.4

# Training utilities
tensorboard==2.15.1
wandb==0.16.1
tqdm==4.66.1

# Configuration
omegaconf==2.3.0
pyyaml==6.0.1

# Web interface
gradio==4.8.0

# Evaluation metrics
torchmetrics==1.2.1
scikit-learn==1.3.2

# Utilities
einops==0.7.0
matplotlib==3.8.2
seaborn==0.13.0

# Development tools
ipykernel==6.27.1
jupyter==1.0.0
black==23.12.0
isort==5.13.2

# Additional dependencies for CLAP
# laion-clap will be installed separately due to dependency conflicts
ftfy==6.1.3
regex==2023.12.25
webdataset==0.2.48

# xFormers for memory-efficient attention (optional)
xformers==0.0.23